refactor: replace BillingDetails and ShippingDetails with a single AddressForm
The two components were ~95% identical — same fields, same layout, same
updateShippingLocation hook. Only differences were the v-model binding name
(billing vs shipping) and isBillingAddressEnabled guards on address fields.

Introduce AddressForm.vue with a showAddressFields prop (default true) that
covers both cases. checkout.vue now passes :show-address-fields="isBillingAddressEnabled"
for billing and omits the prop for shipping. Deletes both old components.
